requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
colorama==0.3.9
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pyserial==3.4
pytest==3.8.0
six==1.11.0
toml==0.9.6
tox==3.3.0
virtualenv==16.0.0
